requirements.txt
# 小白AI图片去水印 - 依赖列表

# 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow>=10.2.0
numpy>=1.24.0
scipy>=1.10.0

# OpenCV（推荐安装）
opencv-python>=4.8.0

# LaMa AI 模型（去水印核心）
torch>=2.0.0
simple-lama-inpainting>=0.1.0

# 打包工具（开发时使用）
# pip install pyinstaller

